refactor(vuln): decompose the scanner god-file by role (WS-D)
Split internal/discovery/vuln/scanner.go (792 LOC) into role-based files
in the same package — pure move: scanner lifecycle/orchestration stays in
scanner.go; standalone version/banner parsing helpers move to parse.go.

All files under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
